Fix remaining mypy errors and flip the CI mypy gate on
Last batch of the mypy triage started with fix-now item #2 (make CI's
mypy check actually block the build instead of `|| true`). Started at
184 pre-existing errors; this brings the count to 0.

Remaining fixes, all pure type-annotation gaps confirmed safe at
runtime (no behaviour change):

- app/templates/aml/factory.py, app/api/routes/governance/controls.py:
  two functions annotated a parameter as `str` while the dict they look
  it up in is keyed by an enum (IndustryType, ControlRiskArea). Both
  enums are `str` mixins, so lookups already worked correctly at
  runtime (str/enum hash+eq match) -- corrected the annotations to the
  real enum types, and passed the enum itself instead of `.value` at
  the one call site where that was cheap to align.

- app/api/routes/security_monitor.py: `_since(days: int)` was called
  with 0.042 (~1 hour) -- timedelta accepts floats fine, so this always
  worked; widened the parameter to `float` to match actual usage.

- app/schemas/transaction_receipt.py, app/services/recommendation_engine.py:
  `getattr(obj, attr, {}).value if hasattr(getattr(obj, attr, None), "value")
  else ...` -- calling getattr(obj, attr, ...) twice per expression
  meant mypy couldn't narrow the type across the hasattr check tied to
  a separate call. Extracted each to a single local variable first, no
  behaviour change, and now type-narrows correctly.

- app/services/monitoring_engine.py: calculate_alert_score() was
  annotated to return dict[str, float], but its `breakdown` dict
  includes weights_used, itself a nested dict -- corrected to
  dict[str, Any], matching how it's actually used (serialized whole
  into a JSON column, never iterated as float-only).

- app/services/examination_pack_service.py (8x), app/services/questionnaire_seed_service.py,
  app/integrations/identity/sumsub.py: unannotated dict literals with
  heterogeneous value types collapsed to `object`/mismatched inference
  the same way the earlier billing/benchmark catalogue dicts did --
  added explicit dict[str, ...] annotations at each.

- app/api/routes/governance/training.py's mypy error (missing stubs for
  dateutil.relativedelta) is fixed by adding types-python-dateutil to
  the CI mypy step, not a source change.

CI change: `pip install types-python-dateutil` added alongside the
existing mypy/types-python-jose install, and `--no-error-summary || true`
removed from the mypy invocation -- the step now fails the build on any
mypy error, same as the ruff and pytest steps already do.

Verified: `mypy app/` reports 0 errors (0 notes aside from three
informational "untyped function body not checked" notes, which are not
errors). `python -c "import app.main"` succeeds. Full test suite: 452
passed, unchanged. ruff clean across app/.

+++ app/api/routes/security_monitor.py
@@ -35,7 +35,7 @@ def _require_super_admin(current_user: User = Depends(_ROLE_GATE)) -> User:
 _PRIV = _require_super_admin
 
 
-def _since(days: int) -> datetime:
+def _since(days: float) -> datetime:
     return datetime.now(timezone.utc) - timedelta(days=days)
